+++ core/services/counselor_service.py
@@ -6,27 +6,58 @@ from core.repositories.counselor_repository import CounselorRepository
 class CounselorService:
 
     @staticmethod
-    def create_counselor(data):
-        
-        user = User.objects.create_user(
-            username=data["username"],
-            password=data["password"]
-        )
+    def create_counselor(request_user, data):
 
-        school = data["school"]
+        if not request_user.is_superuser:
+            raise PermissionError("Only admin can create counselors")
+        
+        if User.objects.filter(username=data["username"]).exists():
+            raise ValueError("Username already exists")
+
+        username = data.pop("username")
+        password = data.pop("password")
+
+        user = User.objects.create_user(
+            username=username,
+            password=password
+        )
 
         return CounselorRepository.create(
             user=user,
-            school=school,
-            full_name=data["full_name"]
+            **data
         )
 
     @staticmethod
-    def update_counselor(counselor_id, data):
+    def update_counselor(request_user, counselor_id, data):
+
+        if not request_user.is_superuser:
+            raise PermissionError("Only admin can update counselors")
+        
         counselor = CounselorRepository.get_by_id(counselor_id)
-        return CounselorRepository.update(counselor, data)
+
+        data.pop("user", None)
+        data.pop("school", None)
+
+        return CounselorRepository.update(counselor, **data)
 
     @staticmethod
-    def delete_counselor(counselor_id):
+    def delete_counselor(request_user, counselor_id):
+
+        if not request_user.is_superuser:
+            raise PermissionError("Only admin can delete counselors")
+
         counselor = CounselorRepository.get_by_id(counselor_id)
-        return CounselorRepository.delete(counselor)
+
+        CounselorRepository.delete(counselor)
+
+    @staticmethod
+    def reset_password(request_user, counselor_id, new_password):
+
+        if not request_user.is_superuser:
+            raise PermissionError("Only admin can reset passwords")
+
+        counselor = CounselorRepository.get_by_id(counselor_id)
+
+        user = counselor.user
+        user.set_password(new_password)
+        user.save()
