+++ core/services/student_event_service.py
@@ -9,9 +9,11 @@ class StudentEventService:
 
         counselor = user.counselor
         school = counselor.school
+        student = data["student"]
 
-        student = Student.objects.for_user(user).get(id=data["student"])
-
+        if student.school != school:
+            raise PermissionError("Student does not belong to your school")
+        
         return StudentEventRepository.create(
             student=student,
             counselor=counselor,
@@ -24,13 +26,23 @@ class StudentEventService:
     @staticmethod
     def update_event(user, event_id, data):
 
-        event = StudentEventRepository.get_by_id(user, event_id)
+        event = StudentEventRepository.get_by_id(event_id)
 
-        return StudentEventRepository.update(event, data)
+        if event.school != user.counselor.school:
+            raise PermissionError("Not allowed")
+        
+        data.pop("school", None)
+        data.pop("counselor", None)
+        data.pop("student", None)
+
+        return StudentEventRepository.update(event, **data)
 
     @staticmethod
     def delete_event(user, event_id):
 
-        event = StudentEventRepository.get_by_id(user, event_id)
+        event = StudentEventRepository.get_by_id(event_id)
+
+        if event.school != user.counselor.school:
+            raise PermissionError("Not allowed")
 
         StudentEventRepository.delete(event)
